Pass key->sk_attno to consistent function in gistindex_keytest.
Right now consistent function is unaware of the key attribute that it is called for.
This in turn makes it impossible to perform any additional processing based on catalog information,
for example partition bounds.

Passing index key attribute number allows the consistent function to lookup
(and possibly cache) catalog information using Relation provided in GISTENTRY.