Teach the planner to pass UniqueKey nodes from a subquery to the parent query.
So far it was only possible to recognize uniqueness of the output of
relatively simple subqueries. With this feature, more complex subqueries can
be checked, even those "hidden" in a view with security_barrier enabled. Thus
there should be more opportunities to apply specific optimizations. Typically,
when the inner relation of join is unique, then INNER join can be replaced
with SEMI join.