Use streaming I/O in BAS to do write-behind.
freelist.c traditionally peformed a sort of one-by-one write-behind, by
cleaning the oldest buffer in the ring at the last moment before
recycling it.

Introduce a more explicit kind of write-behind, that begins once the
consumer of buffers has dirtied a buffer and is ready to unpin it.
Instead of calling ReleaseBuffer(), a new call StrategyReleaseBuffer()
can be used to feed dirty buffers back to the strategy as soon as
possible, so that it can feed them into a write stream.  The write
stream will release the pin some time later, but an interface is
provided at the stratetgy level to order it to drop all pins (eg before
truncating a relation, when you don't want pins hiding in that
pipeline).

The main benefit today is that the write stream generates vectored I/O
up to io_combine_limit and some attempts at avoiding WAL flush stalls.
The write stream also insulates code from future work to support
asynchronous I/O.

The new write stream is enabled only for BAS_BULKWRITE and BAS_VACUUM,
since those strategy types are expected to crash into their dirty data,
but it won't actually do anything yet.  Later commits will add in
StrategyReleaseBuffer() to a few places that can benefit from this
optimization.