Introduce the Asymmetric Join (AJ) feature. Asymmetric Join is an addition to the partitionwise join strategy: it allows the optimiser when it sees join of plain table with partitioned one to attempt joining the table with each of partitions separately.
Here, as usual, we have some pro and cons:
* Smaller JOIN can survive memory allocations or data skews
* Different strategies for different pairs may be more effective.
* It may allow to prune more partitions and with partition constrain select smaller volume of data.
* But of course, here we have increasing complexity of planning.

To implement this feature we add consider_asymmetric_join into the RelOptInfo
structure. It must be true iff consider_partitionwise_join is false.
Implementation follows the logic of partitionwise_join.
We disallow any partitioned relations in the inner of AJ. The key point here
is to resolve the problem with different ways of achieving RelOptInfo which can
need different number of part_rels.
One annoying feature of the implementation is some sort of asymmetry introduced:
we have dependency on the order of incoming relations. If AJ was approved for
a joinrel, possible PWJ for this joinrel with different combination of inner and
outer will be declined.

New GUC enable_asymmetric_join allows to disable this feature, if necessary.

A lot of regression tests were changed or newly created.

TODO:
1. Here we doesn't resolved Ashutosh Bapat's warning on scanning of one
RangeEntryEntry in different parts of the plan. PostgreSQL used to implicitly
assume that each leaf RangeTable (and RelOptInfo) corresponds only one scan node.
For example, such partition pruning code is designed with such assumption.
An qual evaluation code also references index of scan node.
After analysis We didn't found any problems which can be caused by this
inconsistency. Unfortunately, it doesn't guarantee any issues in the future.
2. We should discover on the partition pruning more. May a join order change
impact pruning outcome and how it can influence partitioning schema?