Optimize COPY TO in text format using SIMD
CopyAttributeOutText() does a byte-by-byte loop looking for field delimiters and
escape sequences. Vectorizing this loop using SIMD yields a significant speedup
for wide attributes, assuming that escape sequences are rare.

We don't attempt to apply this optimization when encoding_embeds_ascii is true,
because the bookkeeping required would be complicated.