Use GiST index with mixed integer/float/timestamp widths
A GiST index on a bigint will not be used when compared to an int
literal, unless the user casts the literal to a bigint. This is
surprising and causes invisible slowness. To fix that we can add pg_amop
entries for mixed-width arguments. We do that for all int2/int4/int8
combinations, and also for float4/float8 and date/timestamp/timestamptz
(not that timestamptz is wider than timestamp, but we can follow btree
in allowing mixed operators for them too).

With btree, we have one opfamily for all integer types, and it supports
=(int2,int8), etc. With GiST we have a separate opfamily for each width,
so it's less obvious where to put the mixed-width operators. But it
makes sense for wider opfamilies to include support for smaller types,
so I added =(int2,int8) and =(int4,int8) to gist_int8_ops, and
=(int2,int4) to gist_int4_ops. Also =(float4,float8) to
gist_float8_ops, =(date,timestamptz) and =(timestamp,timestamptz to
gist_timestamptz_ops, and =(date,timestamp) to gist_timestamp_ops. And
as long as we're doing this, we might as well support all five btree
operators (<, <=, =, >=, >).