Introduce new control file parameter 'protected' to define implicit search_path for extension functions.
When enabled, this parameter defines the implicit search_path for
functions and procedures created by extensions if no explicit
search_path is specified. It includes $extension_schema, pg_temp, and
function_schema (if different from the extension's schema). Here
$extension_schema is a special name that dynamically resolves to all
schemas on which the extension depends. This resolution occurs at the
time function or procedure execution.