pg_combinebackup: Detect missing files when possible.
You need a backup_manifest to take an incremental backup, but it's
optional when running pg_combinebackup. In practice, we'll very often
have one, because backup manifests are the default and most people
don't turn them off. And, if we do have one, we can check that no
files mentioned in the manifeste are absent from the corresponding
backup directory.

This falls a long way short of comprehensive verification, because
it only checks against the last backup in the chain, and even then
only for totally missing files. If you want to verify your backups,
it's a much better idea to use pg_verifybackup, which can detect a
much wider variety of problems and which you can run against every
backup you have. But, this check doesn't cost much and has a chance
of alerting you to a very serious problem, so we may as well include
it.

Per discussion with David Steele.

Discussion: http://postgr.es/m/CA+Tgmoa6McCGpaA66dbrGA6onGUytc5Jds_3Y=y7tMaPy1p9mQ@mail.gmail.com