Use the found LLVM version when finding clang
find_program(version:) can be used to restrict the version of the
program to be found. In this case, we need a clang that is compatible
with LLVM.

Prior to 4d8de281b5834c8f5e0be6ae21e884e69dffd4ce, this worked as
intended, but since that change it was possible to pick an incompatible
clang and LLVM combination.

Reported-by: Andres Freund <andres@anarazel.de>