Allow extensions to control join strategy.
At the start of planning, we build a bitmask of allowable join
strategies based on the values of the various enable_* planner GUCs,
indicating which join strategies are allowed. Extensions can override
this mask for an entire subquery using join_search_hook, or, generaly
more usefully, they can change the mask for each call to
add_paths_to_joinrel using a new hook called join_path_setup_hook.
This is sufficient to allow an extension to force the use of
particular join strategies either in general or for specific joins,
and it is also sufficient to allow an extension to force the join
order, including which relation appears on which side of a given join.

There are a number of things that this patch doesn't let you do.
First, it won't help if you want to implement some policy where the
allowable join methods might differ for each individual combination of
input paths (e.g. disable nested loops only when the inner side's
parameterization is not even partially satisfied by the outer side's
paramaeterization). Second, it doesn't allow you to control the
uniquification strategy for a particular joinrel. Third, it doesn't
give you any control over aspects of planning other than join planning.
Future patches may close some of these gaps.