Compress ReorderBuffer spill files using PGLZ
When the content of a large transaction (size exceeding
logical_decoding_work_mem) and its sub-transactions has to be
reordered during logical decoding, then, all the changes are written
on disk in temporary files located in pg_replslot/<slot_name>.

This behavior happens only when the subscriber's option "streaming"
is set to "off", which is the default value. In this case, decoding
of large transactions by multiple replication slots can lead to disk
space saturation and high I/O utilization.

This patch enables data compression/decompression of these temporary
files. Each transaction change that must be written on disk is now
compressed using the in-house compression method PGLZ.