Delay the reset of UNLOGGED relations
This patch set enables the creation of INIT forks within
transactions. Cleanup of such forks after a crash will be managed by
the UNDO log system, which will be introduced in a subsequent
patch. Since the consistency of UNDO logs relies on WAL, operations
involving INIT forks - specifically reinit - must only occur after
recovery has reached consistency.

To prepare for the introduction of the UNDO log system, this commit
adjusts the timing of UNLOGGED relation cleanup. Instead of occurring
before recovery begins, it is now performed either when the
consistency point is reached or at the end of recovery if hot standby
is disabled.