New contrib module: hint_via_alias.
This forces a table to be merge joined, hash joined, or joined via a nested
loop if the table alias starts with mj_, hj_, or nl_, respectively. This
demonstrates that join_path_setup_hook is sufficient to control the join
method.

It forces a table to be sequential scanned, index scanned, index only
scanned, bitmap scanned, or tid scanned if the table alias starts with
ss_, is_, ios_, bs_, o ts_, respectively. This demonstrates that, with
the patch, get_relation_info_hook is sufficient to control the scan
method.

This is not intended for commit.