Add test to cover aclcheck_error_col()
The error message "permission denied for column \"%s\" of relation \"%s\"" is
not in any of the expected output even though it's simple to trigger this error
message. Add a test for the same so as to improve test coverage of this
particular error.

Author: Ashutosh Bapat
Discussion: https://postgr.es/m/CAExHW5swFANiB9JmqRoGg_Rkr%2BM%3Dqh%2Bci_zfOtQXFT%2BA%3D%2BjB-A%40mail.gmail.com