Support regular expressions with nondeterministic collations
For normal regular expressions, the collation doesn't matter, so we
can just allow it.

But for SIMILAR TO, we cannot implement what the SQL standard
requires, so we continue to not allow nondeterministic collations with
SIMILAR TO.  But since SIMILAR TO gets converted to a normal regular
expression, we need to record somehow that the regular expression was
originally a SIMILAR TO.  To do that, we prepend a prefix "***S" which
the RE parsing later recognizes and records.  Based on that, we can
issue the required error message.