pg_combinebackup: Factor some code out of write_reconstructed_file.
The act of copying a single block moves to a new function copy_block(),
and the logic to generate the detailed debugging output is moves to into
a new function debug_reconstruction_plan().  This is preparatory
refactoring for a future patch, but it's not unreasonable on its own
terms, as write_reconstructed_file() has become a fairly lengthy
function.