Allow FDW extensions to support MERGE command via CustomScan
Currently, it is not possible for any fdw extension
to support Merge command, as that's prohibited in the
parser.

In this commit, we allow extensions to support Merge command
via `CustomScan` node by moving the Merge support check
from parser to planner.

For existing fdw, they don't have to change anyting. We change
postgres_fdw mostly for documentation purposes.