Add a warning when using plain text passwords
With CREATE/ALTER ROLE... PASSWORD, plain text passwords may be logged
and can be discovered if someone is listening on the network. We used to
tell users not to use plain text passwords, but it would be better to
warn the user of the issue if the user tries to add/change a password
using a plain text one in the query.

CREATE ROLE and ALTER ROLE now emit warnings when using plain text
passwords. The warnings can be disabled by setting the
plaintext_password_warnings parameter to "off".