Remove the single_copy field from GatherPath.
While single_copy is useful in the Gather plan, it's set regardless the value
in GatherPath: standard_planner() sets the flag on a new instance of Gather,
for which there is no corresponding GatherPath. Furthermore, the field is
never set in GatherPath anyway because parial paths always have
parallel_workers > 0.