JIT compilation per plan node
Introduce a new jit field on Planner struct to make it possible to
decide whether a plan node should be compiled or not.

Previously the cost of the entire query was used to decide whether it
should be compiled or not. This approach has a problem that it can exist
some expressions on plan tree that is not executed frequently, so it is
not worth to compile.

This commit change the consideration of JIT from plan level to
per-plan-node. The plan total cost and the number of times that the node
will be executed is used to decide if the plan node should be compiled
or not.