Introduce planner temporary memory context.
This memory context is dedicated to cutting off memory consumption peaks in
accidentally scaling operations, like huge array estimations or massive clause
estimations caused by a parent expansion.

It includes 'depth' and 'usage' parameters. The first one is used to determine
whether it is really safe to reset the context or whether it is still used
somewhere in the call stack.
Another one counts the number of operations it was involved in. 'Operation'
here means not a memory allocation call but a logical operation, like a single
clause estimation. So, after the end of the operation, If 'depth' = 0,
the optimiser can clean up used memory if the counter is big enough.

We can't track that cleaning memory isn't referenced by an external object.
So, it is the developer's responsibility to put any product of this context
usage in a longer-living memory context.

The first obvious usage of this context is selectivity estimation because
the operation's result is just a floating-point value, and only cached numbers
in the RestrictInfo structure created outside the context may be a by-product
of this operation.