Make pg_interpret_timezone_abbrev() check sp->defaulttype too.
This omission caused it to miss the furthest-back zone abbreviation
when working with timezone data compiled with relatively recent zic
(2018e or newer, I think).  Older versions of zic produced a phony
DST transition at the Big Bang, so that the oldest abbreviation
could always be found in the sp->types[] array; but newer versions
don't do that, so that we must examine defaulttype as well as
the types[] array to be sure of seeing all the abbreviations.

While this has been broken for six or so years, we'd managed not
to notice for two reasons: (1) many platforms are still using
ancient zic for compatibility reasons, so that the issue did not
manifest in builds using --with-system-tzdata; (2) the oldest
zone abbreviation is almost always "LMT", which we weren't
supporting anyway until the current patch series.

While here, update pg_next_dst_boundary() to use sp->defaulttype
as the time type for non-DST zones and times before the oldest
DST transition.  The existing code here predates upstream's
invention of the sp->defaulttype field, and its heuristic for
finding the oldest time type has now been subsumed into the
code that fills in sp->defaulttype.

Possibly this should be back-patched, but I'm unsure whether
there are any visible consequences in released branches.

Per report from Aleksander Alekseev and additional investigation.

Discussion: https://postgr.es/m/CAJ7c6TOATjJqvhnYsui0=CO5XFMF4dvTGH+skzB--jNhqSQu5g@mail.gmail.com